RequestAutomate/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.24
Flask-MySQLdb==0.2.0
requests==2.26.0
click==8.0.3
colorama==0.4.4
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
gunicorn==20.1.0
httpx==0.24.0
